Pin the auto sort order and break the table type import cycle
